Add Player object to game controller to track players more easily.
I found that I needed to keep track of the active player's name and the symbol they were using.  I added a constructor inside the game controller and reworked the play() function to work with the new object.  This also means ascii symbols other than X and O can be used as game pieces.