Add logic to display the state of the game board.
Slight bug with current implementation: the prompt() call fires again before the display is updated for some reason.  putting breaks in the code to debug caused this to work as intended.  It's likely something related to the prompt function.  Issue should be resolved when input method is changed in the future.