Split play function into two separate functions
In the command line form of this game, the loop in the play function would stop and wait until the player typed in their selection.  That solution will not work with the planned implementation of a clickable interface.  In order to resolve the looping problem, I moved the variables in play one level up into gameController and then split play into two functions.  the first function sets up the game board for the first turn.  the second function handles the placement of one piece and moves the game forward 1 turn.